requirements.txt
fastapi
uvicorn
sentence-transformers
torch
Pillow
jinja2
python-multipart
